+++ src/steps/resource-manager/monitor/converters.ts
@@ -10,6 +10,7 @@ import {
   LogSettings,
   MetricSettings,
 } from '@azure/arm-monitor/esm/models';
+import { normalizeId } from '../utils/normalizeId';
 
 export function createMonitorLogProfileEntity(
   webLinker: AzureWebLinker,
@@ -80,16 +81,6 @@ function createSettingEntity(
       },
     },
   });
-}
-
-/**
- * When getting Diagnostic Settings for some Azure Resources, the Diagnostic Settings id is returned from the client without the leading slash.
- * This function adds the leading slash to an id, if it does not exist
- * @param id The resource URI of an Azure resource
- */
-function normalizeId(id: string | undefined): string | undefined {
-  if (!id) return id;
-  return id.startsWith('/') ? id : `/${id}`;
 }
 
 export function createDiagnosticLogSettingEntity(